+++ node/src/chain_spec.rs
@@ -1,10 +1,11 @@
 use cumulus_primitives_core::ParaId;
+use runtime::{AccountId, AuraId, Signature, EXISTENTIAL_DEPOSIT};
 use sc_chain_spec::{ChainSpecExtension, ChainSpecGroup};
 use sc_service::ChainType;
 use serde::{Deserialize, Serialize};
 use sp_core::{sr25519, Pair, Public};
 use sp_runtime::traits::{IdentifyAccount, Verify};
-use storage_hub_runtime::{AccountId, AuraId, Signature, EXISTENTIAL_DEPOSIT};
+use storage_hub_runtime as runtime;
 
 /// Specialized `ChainSpec` for the normal parachain runtime.
 pub type ChainSpec = sc_service::GenericChainSpec<(), Extensions>;
@@ -56,8 +57,8 @@ where
 /// Generate the session keys from individual elements.
 ///
 /// The input must be a tuple of individual keys (a single arg for now since we have just one key).
-pub fn template_session_keys(keys: AuraId) -> storage_hub_runtime::SessionKeys {
-    storage_hub_runtime::SessionKeys { aura: keys }
+pub fn template_session_keys(keys: AuraId) -> runtime::SessionKeys {
+    runtime::SessionKeys { aura: keys }
 }
 
 pub fn development_config() -> ChainSpec {
@@ -68,7 +69,7 @@ pub fn development_config() -> ChainSpec {
     properties.insert("ss58Format".into(), 42.into());
 
     ChainSpec::builder(
-        storage_hub_runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
+        runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
         Extensions {
             relay_chain: "rococo-local".into(),
             // You MUST set this to the correct network!
@@ -119,7 +120,7 @@ pub fn local_testnet_config() -> ChainSpec {
 
     #[allow(deprecated)]
     ChainSpec::builder(
-        storage_hub_runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
+        runtime::WASM_BINARY.expect("WASM binary was not built, please build it!"),
         Extensions {
             relay_chain: "rococo-local".into(),
             // You MUST set this to the correct network!
