+++ node/src/cli.rs
@@ -24,8 +24,11 @@ pub enum Subcommand {
     /// Remove the whole chain.
     PurgeChain(cumulus_client_cli::PurgeChainCmd),
 
-    /// Export the genesis state of the parachain.
-    ExportGenesisState(cumulus_client_cli::ExportGenesisHeadCommand),
+    /// Export the genesis head data of the parachain.
+    ///
+    /// Head data is the encoded block header.
+    #[command(alias = "export-genesis-state")]
+    ExportGenesisHead(cumulus_client_cli::ExportGenesisHeadCommand),
 
     /// Export the genesis wasm of the parachain.
     ExportGenesisWasm(cumulus_client_cli::ExportGenesisWasmCommand),
@@ -37,7 +40,7 @@ pub enum Subcommand {
 
     /// Try-runtime has migrated to a standalone
     /// [CLI](<https://github.com/paritytech/try-runtime-cli>). The subcommand exists as a stub and
-    /// deprecation notice. It will be removed entirely some time after Janurary 2024.
+    /// deprecation notice. It will be removed entirely some time after January 2024.
     TryRuntime,
 }
