+++ pallets/file-system/Cargo.toml
@@ -52,7 +52,7 @@ runtime-benchmarks = [
 ]
 std = [
     "codec/std",
-    "frame-benchmarking/std",
+    "frame-benchmarking?/std",
     "frame-support/std",
     "frame-system/std",
     "pallet-storage-providers/std",
