+++ pallets/proofs-dealer/Cargo.toml
@@ -53,7 +53,7 @@ runtime-benchmarks = [
 ]
 std = [
     "codec/std",
-    "frame-benchmarking/std",
+    "frame-benchmarking?/std",
     "frame-support/std",
     "frame-system/std",
     "pallet-balances/std",
