+++ runtime/Cargo.toml
@@ -77,6 +77,7 @@ cumulus-pallet-xcmp-queue = { workspace = true }
 cumulus-primitives-aura = { workspace = true }
 cumulus-primitives-core = { workspace = true }
 cumulus-primitives-utility = { workspace = true }
+cumulus-primitives-storage-weight-reclaim = { workspace = true }
 pallet-collator-selection = { workspace = true }
 parachains-common = { workspace = true }
 parachain-info = { workspace = true }
@@ -94,6 +95,7 @@ std = [
     "cumulus-primitives-aura/std",
     "cumulus-primitives-core/std",
     "cumulus-primitives-utility/std",
+    "cumulus-primitives-storage-weight-reclaim/std",
     "frame-benchmarking?/std",
     "frame-executive/std",
     "frame-support/std",
