+++ support/traits/Cargo.toml
@@ -41,7 +41,7 @@ runtime-benchmarks = [
 ]
 std = [
     "codec/std",
-    "frame-benchmarking/std",
+    "frame-benchmarking?/std",
     "frame-support/std",
     "frame-system/std",
     "scale-info/std",
